src/saaga_mcp_servers/utils/mcp-researcher/pyproject.toml
[project]
name = "mcp-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andrew Oseen", email = "amovfx@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "gpt-researcher>=0.14.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "mcp>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
mcp-researcher = "mcp_researcher.cli.main:main"
mcp-researcher-server = "mcp_researcher.mcp.server:main"
mcp-researcher-config = "mcp_researcher.cli.config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
